+++ frontend/index.html
@@ -10,8 +10,6 @@
     <meta name="mobile-web-app-capable" content="yes">
     <meta name="apple-mobile-web-app-capable" content="yes">
     <title>Jellyfin Vue</title>
-    <script type="module" src="src/splashscreen.ts" fetchpriority="high"></script>
-    <script type="module" src="src/main.ts"></script>
   </head>
   <body class="j-splash">
     <img src="./icon.svg" alt="Jellyfin Logo" loading="eager" decoding="sync">
