+++ frontend/scripts/paths.ts
@@ -1,3 +1,3 @@
 import { resolve } from 'node:path';
 
-export const localeFilesFolder = resolve('locales/**');
+export const localeFilesFolder = resolve(import.meta.dirname, '../locales/**');
