+++ packages/vite-plugins/package.json
@@ -11,6 +11,7 @@
   },
   "devDependencies": {
     "@jellyfin-vue/configs": "*",
+    "find-up-simple": "1.0.0",
     "pretty-bytes": "6.1.1",
     "sonda": "0.7.1",
     "vite": "6.0.7"
